feat: add rara-plugins crate with Claude Code plugin support
- New crate crates/rara-plugins with loader + exec + types
- Plugin discovery from directory with plugin.json + hooks.json parsing
- Command hook execution (spawn shell + stdin JSON + exit code)
- 6 lifecycle events: Stop, Pre/PostToolUse, UserPromptSubmit, SessionStart/End
- Spec at docs/features/claude-plugin-runtime.md
- Removed old src/plugin_loader.rs and src/plugin_exec.rs
