+++ crates/rara-plugins/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "rara-plugins"
+version = "0.0.1"
+edition = "2024"
+license = "Apache-2.0"
+description = "Claude Code plugin loader and command hook executor for RARA"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+tokio = { version = "1", features = ["process", "io-util", "time", "rt"] }
+tempfile = "3"
